deplacer getAvancementStatistique vers domain chantiers
