+++ src/server/domain/chantier/ChantierRepository.interface.ts
@@ -3,10 +3,7 @@ import {
   chantier_territoire as PrismaChantierTerritoire,
   chantier_territoire_jalon as PrismaChantierTerritoireJalon,
 } from "@prisma/client";
-import Chantier from "@/server/domain/chantier/Chantier.interface";
-import { Maille } from "@/server/domain/maille/Maille.interface";
 import { Habilitations } from "@/server/domain/utilisateur/habilitation/Habilitation.interface";
-import { AvancementsStatistiques } from "@/components/_commons/Avancements/Avancements.interface";
 import { ProfilCode } from "@/server/domain/utilisateur/Utilisateur.interface";
 import { RepartitionMeteoChantiers } from "@/server/chantiers/domain/RepartitionMeteoChantiers";
 import { FiltreQueryParams } from "@/server/chantiers/app/contrats/FiltreQueryParams";
@@ -25,12 +22,6 @@ export default interface ChantierRepository {
       })[];
     }
   >;
-  getChantierStatistiques(
-    habilitations: Habilitations,
-    listeChantier: Chantier["id"][],
-    maille: Maille,
-    jalon: number,
-  ): Promise<AvancementsStatistiques>;
   recupererLaRepartitionMeteo(
     chantiersLectureIds: string[],
     territoireCode: string,
