+++ src/server/infrastructure/accès_données/chantier/ChantierSQLRepository.integration.test.ts
@@ -256,441 +256,6 @@ describe("ChantierSQLRepository", () => {
     });
   });
 
-  describe("#getChantierStatistiques", () => {
-    it("quand le nombre de territoire demandé est pair, doit récupérer les statistiques d'une liste de chantier", async () => {
-      // Given
-      const listeChantierIds = ["CH-001", "CH-002"];
-
-      const habilitation = {
-        lecture: {
-          chantiers: ["CH-001", "CH-002", "CH-003", "CH-004"],
-          territoires: ["NAT-FR"],
-        },
-      } as unknown as Utilisateur["habilitations"];
-
-      await prisma.chantier_identite.createMany({
-        data: [
-          {
-            id: "CH-001",
-            nom: "Chantier 001",
-          },
-          {
-            id: "CH-002",
-            nom: "Chantier 002",
-          },
-          {
-            id: "CH-003",
-            nom: "Chantier 003",
-          },
-        ],
-      });
-
-      await prisma.chantier_territoire.createMany({
-        data: [
-          {
-            id: "CH-001",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "SOLEIL",
-            territoire_code: "DEPT-87",
-          },
-          {
-            id: "CH-001",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            meteo: "COUVERT",
-            territoire_code: "DEPT-88",
-          },
-          {
-            id: "CH-002",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "NON_RENSEIGNEE",
-            territoire_code: "DEPT-87",
-          },
-          {
-            id: "CH-002",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            meteo: "COUVERT",
-            territoire_code: "DEPT-88",
-          },
-          {
-            id: "CH-002",
-            zone_id: "RO1",
-            maille: "REG",
-            code_insee: "01",
-            meteo: "COUVERT",
-            territoire_code: "REG-01",
-          },
-          {
-            id: "CH-003",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "NON_RENSEIGNEE",
-            territoire_code: "DEPT-87",
-          },
-        ],
-      });
-
-      await prisma.chantier_territoire_jalon.createMany({
-        data: [
-          {
-            id: "CH-001",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 10,
-          },
-          {
-            id: "CH-001",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            territoire_code: "DEPT-88",
-            jalon: 2025,
-            taux_avancement: 12,
-          },
-          {
-            id: "CH-002",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 14,
-          },
-          {
-            id: "CH-002",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            territoire_code: "DEPT-88",
-            jalon: 2025,
-            taux_avancement: 16,
-          },
-          {
-            id: "CH-002",
-            zone_id: "RO1",
-            maille: "REG",
-            code_insee: "01",
-            territoire_code: "REG-01",
-            jalon: 2025,
-            taux_avancement: 50,
-          },
-          {
-            id: "CH-003",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 18,
-          },
-        ],
-      });
-
-      // When
-      const result = await prismaChantierRepository.getChantierStatistiques(
-        habilitation,
-        listeChantierIds,
-        "departementale",
-        2025,
-      );
-
-      // Then
-      expect(result).toEqual({
-        médiane: 13,
-        maximum: 14,
-        minimum: 12,
-      });
-    });
-
-    it("quand le nombre de territoire demandé est impair, doit récupérer les statistiques d'une liste de chantier", async () => {
-      // Given
-      const listeChantierIds = ["CH-001", "CH-002"];
-
-      const habilitation = {
-        lecture: {
-          chantiers: ["CH-001", "CH-002", "CH-003", "CH-004"],
-          territoires: ["NAT-FR"],
-        },
-      } as unknown as Utilisateur["habilitations"];
-
-      await prisma.chantier_identite.createMany({
-        data: [
-          {
-            id: "CH-001",
-            nom: "Chantier 001",
-          },
-          {
-            id: "CH-002",
-            nom: "Chantier 002",
-          },
-          {
-            id: "CH-003",
-            nom: "Chantier 003",
-          },
-        ],
-      });
-
-      await prisma.chantier_territoire.createMany({
-        data: [
-          {
-            id: "CH-001",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "SOLEIL",
-            territoire_code: "DEPT-87",
-          },
-          {
-            id: "CH-001",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            meteo: "COUVERT",
-            territoire_code: "DEPT-88",
-          },
-          {
-            id: "CH-001",
-            zone_id: "D89",
-            maille: "DEPT",
-            code_insee: "89",
-            meteo: "COUVERT",
-            territoire_code: "DEPT-89",
-          },
-          {
-            id: "CH-002",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "NON_RENSEIGNEE",
-            territoire_code: "DEPT-87",
-          },
-          {
-            id: "CH-002",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            meteo: "COUVERT",
-            territoire_code: "DEPT-88",
-          },
-          {
-            id: "CH-002",
-            zone_id: "RO1",
-            maille: "REG",
-            code_insee: "01",
-            meteo: "COUVERT",
-            territoire_code: "REG-01",
-          },
-          {
-            id: "CH-003",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "NON_RENSEIGNEE",
-            territoire_code: "DEPT-87",
-          },
-        ],
-      });
-
-      await prisma.chantier_territoire_jalon.createMany({
-        data: [
-          {
-            id: "CH-001",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 10,
-          },
-          {
-            id: "CH-001",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            territoire_code: "DEPT-88",
-            jalon: 2025,
-            taux_avancement: 12,
-          },
-          {
-            id: "CH-001",
-            zone_id: "D89",
-            maille: "DEPT",
-            code_insee: "89",
-            territoire_code: "DEPT-89",
-            jalon: 2025,
-            taux_avancement: 22,
-          },
-          {
-            id: "CH-002",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 14,
-          },
-          {
-            id: "CH-002",
-            zone_id: "D88",
-            maille: "DEPT",
-            code_insee: "88",
-            territoire_code: "DEPT-88",
-            jalon: 2025,
-            taux_avancement: 16,
-          },
-          {
-            id: "CH-002",
-            zone_id: "RO1",
-            maille: "REG",
-            code_insee: "01",
-            territoire_code: "REG-01",
-            jalon: 2025,
-            taux_avancement: 50,
-          },
-          {
-            id: "CH-003",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 18,
-          },
-        ],
-      });
-
-      // When
-      const result = await prismaChantierRepository.getChantierStatistiques(
-        habilitation,
-        listeChantierIds,
-        "departementale",
-        2025,
-      );
-
-      // Then
-      expect(result).toEqual({
-        médiane: 14,
-        maximum: 22,
-        minimum: 12,
-      });
-    });
-
-    it("calcule les statistiques pour le jalon sélectionné", async () => {
-      // Given
-      const listeChantierIds = ["CH-001", "CH-002"];
-
-      const habilitation = {
-        lecture: {
-          chantiers: ["CH-001", "CH-002"],
-          territoires: ["NAT-FR"],
-        },
-      } as unknown as Utilisateur["habilitations"];
-
-      await prisma.chantier_identite.createMany({
-        data: [
-          {
-            id: "CH-001",
-            nom: "Chantier 001",
-          },
-          {
-            id: "CH-002",
-            nom: "Chantier 002",
-          },
-        ],
-      });
-
-      await prisma.chantier_territoire.createMany({
-        data: [
-          {
-            id: "CH-001",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "SOLEIL",
-            territoire_code: "DEPT-87",
-          },
-          {
-            id: "CH-002",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            meteo: "COUVERT",
-            territoire_code: "DEPT-87",
-          },
-        ],
-      });
-
-      await prisma.chantier_territoire_jalon.createMany({
-        data: [
-          // Jalon 2025
-          {
-            id: "CH-001",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 10,
-          },
-          {
-            id: "CH-002",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2025,
-            taux_avancement: 20,
-          },
-          // Jalon 2026 - valeurs différentes qui ne doivent pas être prises en compte
-          {
-            id: "CH-001",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2026,
-            taux_avancement: 80,
-          },
-          {
-            id: "CH-002",
-            zone_id: "D87",
-            maille: "DEPT",
-            code_insee: "87",
-            territoire_code: "DEPT-87",
-            jalon: 2026,
-            taux_avancement: 90,
-          },
-        ],
-      });
-
-      // When
-      const result = await prismaChantierRepository.getChantierStatistiques(
-        habilitation,
-        listeChantierIds,
-        "departementale",
-        2025,
-      );
-
-      // Then
-      expect(result).toEqual({
-        médiane: 15,
-        minimum: 15,
-        maximum: 15,
-      });
-    });
-  });
-
   describe("#recupererLaRepartitionMeteo", () => {
     it("quand on a l'option estBarometre et territorialisation à ['regionale', 'departementale'], doit remonter la répartition météo des chantiers qui sont du barometre et territorialisés", async () => {
       // Given
